requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
unstructured
pypdf
docx2txt
tiktoken